requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Authlib==0.15.5
bcrypt==3.2.2
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.1
cryptography==3.2
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
google==3.0.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-cloud-translate==3.7.2
google-cloud-vision==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==2.9
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
paramiko==2.10.4
proto-plus==1.20.3
protobuf==3.20.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.14.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.25.8
websocket-client==0.59.0
Werkzeug==1.0.1
zipp==3.8.0
